merge master

+++ src/Form/Field/File.php
@@ -9,6 +9,7 @@ use Symfony\Component\HttpFoundation\File\UploadedFile;
 class File extends Field
 {
     use UploadField;
+    use HasValuePicker;
 
     /**
      * Css.
@@ -101,6 +102,10 @@ class File extends Field
      */
     public function prepare($file)
     {
+        if ($this->picker) {
+            return parent::prepare($file);
+        }
+
         if (request()->has(static::FILE_DELETE_FLAG)) {
             return $this->destroy();
         }
@@ -145,6 +150,18 @@ class File extends Field
     }
 
     /**
+     * Hides the file preview.
+     *
+     * @return $this
+     */
+    public function hidePreview()
+    {
+        return $this->options([
+            'showPreview' => false,
+        ]);
+    }
+
+    /**
      * Initialize the caption.
      *
      * @param string $caption
@@ -186,11 +203,11 @@ EOT;
 
             $this->script .= <<<EOT
 $("input{$this->getElementClassSelector()}").on('filebeforedelete', function() {
-    
+
     return new Promise(function(resolve, reject) {
-    
+
         var remove = resolve;
-    
+
         swal({
             title: "{$text['title']}",
             type: "warning",
@@ -218,6 +235,10 @@ EOT;
      */
     public function render()
     {
+        if ($this->picker) {
+            return $this->renderFilePicker();
+        }
+
         $this->options(['overwriteInitial' => true, 'msgPlaceholder' => trans('admin.choose_file')]);
 
         $this->setupDefaultOptions();
